test(ui): move the arrow tests beside their sources and name the marker requirement in docs
The caret and focus-left tests join focuskeys_test, right-step joins
keys_test and the beta notice joins notices_test, matching the
test-beside-source convention. The usage table now says the left step
out of focus needs the tool's activity_cutoff marker on the caret's row.
