+++ docs/usage.md
@@ -20,7 +20,7 @@ Agent sessions live on a private tmux server named `agentmgr`, so they never mix
 | `A` | Attach session full screen (Settings can swap it with `enter`) |
 | `ctrl+q` | Inside a session: back to the manager |
 | `→` | Step into the row: focus the session, or open the group. In beta; Settings (`s`) can turn the pair off |
-| `←` | Step out: close the group, or — focused, with the caret at the start of the agent's prompt — back to the manager. Anywhere else in the prompt it moves the caret as usual |
+| `←` | Step out: close the group, or — focused, with the caret at the start of the agent's prompt — back to the manager. This needs the tool's prompt marker (its `activity_cutoff`) on the caret's row, so a CLI without one keeps `←` entirely; anywhere else in the prompt it moves the caret as usual |
 | `K` / `J` (or `shift+↑` / `shift+↓`) | Reorder session or group among its visible siblings |
 | `m` | Move session to another group |
 | `r` | Rename session / edit tool; edit group name and default path |
